+++ integrations/fastify/src/plugin/interface.ts
@@ -29,15 +29,18 @@ type EjectInterfacePluginOptions = Parameters<
   EjectInterfaceAPI["api"]["post"]
 >[0] & {
   ejectHost?: string; // defaults to EJECT_HOST or localhost:3000
-  specOutput?: string | false | null; // defaults to ./eject-openapi-spec.3.1.0.json
 };
+
+const refStringTransformation = (ref: string) =>
+  ref.startsWith("#/components/schemas/")
+    ? ref
+    : `#/components/schemas/${ref.replace("#", "")}`;
 
 const EjectInterfacePluginCallback: FastifyPluginAsync<
   EjectInterfacePluginOptions
 > = async (fastify, options) => {
   const {
-    ejectHost = process.env.EJECT_HOST || "http://localhost:3000",
-    specOutput = "eject-openapi-spec.3.1.0.json",
+    ejectHost = process.env.EJECT_HOST || "http://localhost:3734",
     ...apiOptions
   } = options;
 
@@ -60,7 +63,6 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
   // Use this hook to fire off our schemas etc
   fastify.addHook("onRoute", (routeOptions: EjectRouteOption) => {
     routes.push(routeOptions);
-    console.log("Route registered: ", routeOptions.url, routeOptions.method);
   });
 
   // Use this hook to compile schemas once the API starts listening
@@ -75,10 +77,7 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
         const [schemaKey, schemaValue] = schema;
         await interfaceApi.components.post(key, "schema", {
           name: schemaKey.replace("#", ""),
-          component: transformRefs(
-            schemaValue,
-            (ref) => `#/components/schemas/${ref.replace("#", "")}`
-          ),
+          component: transformRefs(schemaValue, refStringTransformation),
         });
       }
 
@@ -89,14 +88,7 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
           in: target,
           required: value.required || target === "path" ? true : false,
           description: value?.description,
-          content: {
-            "application/json": {
-              schema: transformRefs(
-                value,
-                (ref) => `#/components/schemas/${ref.replace("#", "")}`
-              ),
-            },
-          },
+          schema: transformRefs(value, refStringTransformation),
         });
 
       for await (const route of routes) {
@@ -134,8 +126,7 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
                         "application/json": {
                           schema: transformRefs(
                             route.schema?.body,
-                            (ref) =>
-                              `#/components/schemas/${ref.replace("#", "")}`
+                            refStringTransformation
                           ),
                         },
                       },
@@ -152,8 +143,7 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
                           "application/json": {
                             schema: transformRefs(
                               value,
-                              (ref) =>
-                                `#/components/schemas/${ref.replace("#", "")}`
+                              refStringTransformation
                             ),
                           },
                         },
@@ -167,12 +157,7 @@ const EjectInterfacePluginCallback: FastifyPluginAsync<
         }
       }
 
-      if (specOutput) {
-        await fs.writeFile(
-          specOutput,
-          JSON.stringify(await interfaceApi.api.get(key), undefined, 2)
-        );
-      }
+      await interfaceApi.close();
     }, 2500);
   });
 };