+++ packages/cli/src/routes/api/[id]/route/route.get.ts
@@ -1,8 +1,0 @@
-import { Api } from "../../../../api.js";
-import type { RouteHandler, RouteInfo } from "@eject/fastify";
-import { Type } from "@sinclair/typebox";
-
-export default (async (
-  api: Api,
-  details: RouteInfo
-) => {}) satisfies RouteHandler;