+++ starters/typescript/src/index.ts
@@ -1,23 +1,19 @@
-import path from "path";
+import { fileURLToPath } from "node:url";
+import path from "node:path";
 import * as eject from "@eject/fastify";
 import { api } from "./api.js";
 
-// Build an API
-const start = async () => {
-  await api.register(eject.ejectInterface);
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-  // Register Eject routes plugin
-  await api.register(eject.routes, {
-    // prefix: "/api",
-    dir: path.join(__dirname, "routes"),
-  });
+// Register Eject routes plugin
+await api.register(eject.routes, {
+  // prefix: "/api",
+  dir: path.join(__dirname, "routes"),
+});
 
-  try {
-    await api.listen({ port: parseInt(process.env.PORT || "3000") });
-  } catch (err) {
-    api.log.error(err);
-    process.exit(1);
-  }
-};
-
-start();
+try {
+  await api.listen({ port: parseInt(process.env.PORT || "3000") });
+} catch (err) {
+  api.log.error(err);
+  process.exit(1);
+}