+++ integrations/javascript/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @eject/javascript-sdk
 
+## 1.3.0
+
+### Minor Changes
+
+- 0fdbebc: CI fix for stdin input
+
+### Patch Changes
+
+- Updated dependencies [0fdbebc]
+  - @eject/interface@1.7.0
+
+## 1.2.0
+
+### Minor Changes
+
+- fd04e4b: Added initial CLI support for Ejection
+
+### Patch Changes
+
+- Updated dependencies [fd04e4b]
+  - @eject/interface@1.6.0
+
 ## 1.1.0
 
 ### Minor Changes