+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eject/cli
 
+## 1.3.0
+
+### Minor Changes
+
+- 0fdbebc: CI fix for stdin input
+
+## 1.2.0
+
+### Minor Changes
+
+- fd04e4b: Added initial CLI support for Ejection
+
 ## 1.1.0
 
 ### Minor Changes