+++ packages/interface/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eject/interface
 
+## 1.7.0
+
+### Minor Changes
+
+- 0fdbebc: CI fix for stdin input
+
+## 1.6.0
+
+### Minor Changes
+
+- fd04e4b: Added initial CLI support for Ejection
+
 ## 1.5.0
 
 ### Minor Changes